+++ rigsys/modules/export/unreal_translation_logic/utility_modules/control_shape_override.py
@@ -1,0 +1,9 @@
+"""Translator for utility.ControlShapeOverride module."""
+
+from ..workflow_core import run_module_translation
+
+
+def translate(control_rig_bp, module_payload, shared):
+    """Translate ControlShapeOverride utility payload."""
+    _ = control_rig_bp
+    return run_module_translation(module_payload, shared)